[general] changed all GetWindowLong(...,GW_STYLE) to GetWindowStyle() macro. changed all GetWindowLong(...,GW_EXSTYLE) to GetWindowExStyle() macro. commented out or removed unused vars. changed all SetWindowLongPtr(...,GWLP_WNDPROC,...) to SubclassWindow() macro. changed all windows style vars to be DWORD's instead of long's changed all the DcxXXXXXX::getStyles() to use TString styles(__super::getStyles()) removed some unreachable code: `else return ...` followed by a `return` mostly. changed several wrongly used temp TString object as args, when they should have been assigned to an object first. added flag CTLF_ALLOW_ALL for use with the controlFactory() when all controls are allowed, instead of -1. added flag DCX_EVENT_ALL to signify that all events are allowed instead of using -1. changed all vars for alpha values to BYTE's changed all functions that return alpha values to return BYTE's changed all calls to GetGValue() macro to pre-mask the colour as a word value. (doesn't change result just for compilers peace of mind) several vars gave `possible usage of uninitialized vars` warnings, fixed that. changed all COLORREF's to use CLR_INVALID instead of -1. changed several TString.gettok(...," ") to TString.gettok(....) as the space is the default token. added comment about p_Img->GetHICON() causing exception warnings.
[comboex]
changed matchItemText() to take const args.

[listview]
added /xdid -S to perform the mass save of items to hashtable/file/@window, NOT USABLE yet.
changed autoSize() to take an int iFlags instead of UINT as we are looking for negative values.
changed parseHeaderFlags2() to return an INT for the same reason.
changed /xdid -w to check for icon var being NULL before destroying it.

[dcxml]
changed xdialogEX() & xdidEX() to use the TString.vprintf() function.

[progressbar][updown]
added an error report when tooltips style is used but the tooltips window fails to initialize.

[rebar]
changed getControl() to take a const arg.

[toolbar]
changed /xdid -a to mask the buttonStyles var as a byte when setting fsStyle.

[treeview]
changed getPathFromItem() to do a specific != NULL check in the while loop.

[webctrl]
added m_bHideEvents var thats true by default, untill about:blank is loaded. This hides all web events untill the loading is finished.
changed the `Created Browser Window!!!` message to only appear in debug builds

[dock]
changed DcxDock::DcxDock() to take a const type args.

[xpopup]
fixed background not being drawn when using a custom style & no bitmap supplied.
fixed /xpopup -s when relative to a window the pos was incorrect.


git-svn-id: http://dcx.googlecode.com/svn@991 17a06da3-8c20-fa79-550f-e9588e724b77

+++ Classes/dcxtoolbar.h
@@ -123,7 +123,7 @@ protected:
 	int getFreeButtonID( ) const;
 	int getIndexToCommand( const int iIndex ) const;
 
-	static UINT parseButtonStateFlags( const TString & flags );
+	static BYTE parseButtonStateFlags( const TString & flags );
 	static UINT parseButtonStyleFlags( const TString & flags );
 
 	void resetContent( );
